Add unit-test for customize-snapshot class
